requirements.txt
opencv-python
numpy<2
face_recognition
mediapipe
pyserial
